+++ src/pyuvdata/uvdata/mwa_corr_fits.py
@@ -154,6 +154,18 @@ def read_metafits(
         else:
             pass
 
+        # deripple on = 1
+        deripple = meta_hdr["DERIPPLE"]
+
+        # turn off pfb correction if it was corrected in the correlator
+        if deripple == 1:
+            warnings.warn(
+                "No coarse band shape will be removed from this data "
+                "because DERIPPLE is on."
+            )
+        else:
+            pass
+
         # frequency averaging factor
         avg_factor = meta_hdr["NAV_FREQ"]
 
@@ -215,6 +227,7 @@ def read_metafits(
         "flagged_ant_inds": flagged_ant_inds,
         "int_time": int_time,
         "start_flag": start_flag,
+        "deripple": deripple,
         "obs_freq_center": obs_freq_center,
         "avg_factor": avg_factor,
         "coarse_chans": coarse_chans,
@@ -1308,6 +1321,7 @@ class MWACorrFITS(UVData):
         ant_2_inds,
         cb_array,
         dig_gains,
+        deripple,
         nsamples,
         num_fine_chans,
         correct_van_vleck,
@@ -1366,7 +1380,7 @@ class MWACorrFITS(UVData):
             cb_data /= dig_gains1
             cb_data /= dig_gains2
         # remove coarse band
-        if remove_coarse_band:
+        if remove_coarse_band and deripple != 1:
             if freq_inds_dict is not None:
                 cb_data /= cb_array[freq_inds_dict[cb_num]["coarse_inds"], np.newaxis]
             else:
@@ -1386,6 +1400,7 @@ class MWACorrFITS(UVData):
         ant_2_inds,
         avg_factor,
         dig_gains,
+        deripple,
         spw_inds,
         num_fine_chans,
         flagged_ant_inds,
@@ -1468,7 +1483,7 @@ class MWACorrFITS(UVData):
         else:
             dig_gains = None
         # get pfb response shape
-        if remove_coarse_band:
+        if remove_coarse_band and deripple != 1:
             self.history += " Divided out pfb coarse channel bandpass."
             cb_array = self._get_pfb_shape(avg_factor, mwax)
         else:
@@ -1482,6 +1497,7 @@ class MWACorrFITS(UVData):
                 ant_2_inds,
                 cb_array,
                 dig_gains,
+                deripple,
                 nsamples,
                 num_fine_chans,
                 correct_van_vleck,
@@ -2150,6 +2166,7 @@ class MWACorrFITS(UVData):
                     ant_2_inds,
                     meta_dict["avg_factor"],
                     meta_dict["dig_gains"],
+                    meta_dict["deripple"],
                     orig_spw_inds,
                     num_fine_chans,
                     meta_dict["flagged_ant_inds"],